requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-dotenv==1.0.0
ai21==3.1.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
requests==2.28.2
replicate==0.20.0
